Address comments from MaskRay #2
